Updated the engine to 1.7.0, some code cleanup, started work on issue #8

+++ Classifier.Interfaces/Helpers/LogEntry.cs
@@ -27,6 +27,6 @@
     /// <summary>
     /// The unix date/time of the log entry
     /// </summary>
-    public int Unixtime { get; set; }
+    public long Unixtime { get; set; }
   }
 }
