+++ Classifier.Interop/ClassifyEngine.h
@@ -58,7 +58,7 @@ public:
 
   // Log
   int Log(String^ source, String^ entry);
-  bool ClearLogEntries(int olderThan);
+  bool ClearLogEntries(long long olderThan);
   int GetLogEntries(List<Classifier::Interfaces::Helpers::LogEntry^> ^% entries, int max);
 
 protected:
@@ -103,7 +103,7 @@ protected:
   typedef int(__stdcall *f_GetVersion)();
 
   typedef int(__stdcall *f_Log)(const char16_t*, const char16_t*);
-  typedef bool(__stdcall *f_ClearLogEntries)(int);
+  typedef bool(__stdcall *f_ClearLogEntries)(long long);
 
   struct LogEntry
   {
