+++ Classifier.Outlook/interfaces/ILogger.cs
@@ -65,6 +65,6 @@ namespace myoddweb.classifier.interfaces
     /// </summary>
     /// <param name="olderThan"> the date we want to delte.</param>
     /// <returns>success or not</returns>
-    bool ClearLogEntries(int olderThan);
+    bool ClearLogEntries(long olderThan);
   }
 }
